feat(auth): added auth routes and jwt protected

+++ src/routes/memories.ts
@@ -3,8 +3,13 @@ import { prisma } from '../lib/prisma'
 import { z } from 'zod'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
+
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
+      where: { userId: request.user.sub },
       orderBy: { createdAt: 'asc' },
     })
     return memories.map((memory) => ({
@@ -15,12 +20,16 @@ export async function memoriesRoutes(app: FastifyInstance) {
     }))
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, response) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
     const { id } = paramsSchema.parse(request.params)
-    return await prisma.memory.findUniqueOrThrow({ where: { id } })
+    const memory = await prisma.memory.findUniqueOrThrow({ where: { id } })
+    if (!memory.isPublic && request.user.sub !== memory.userId) {
+      return response.status(401).send()
+    }
+    return memory
   })
 
   app.post('/memories', async (request) => {
@@ -36,12 +45,12 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '4ec7018f-541b-4d27-82f6-7c209b1f340b',
+        userId: request.user.sub,
       },
     })
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, response) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -52,17 +61,27 @@ export async function memoriesRoutes(app: FastifyInstance) {
       isPublic: z.coerce.boolean().default(false),
     })
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
+
+    const memory = await prisma.memory.findUniqueOrThrow({ where: { id } })
+    if (request.user.sub !== memory.userId) {
+      return response.status(401).send()
+    }
+
     return await prisma.memory.update({
       where: { id },
       data: { content, coverUrl, isPublic },
     })
   })
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, response) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
     const { id } = paramsSchema.parse(request.params)
+    const memory = await prisma.memory.findUniqueOrThrow({ where: { id } })
+    if (request.user.sub !== memory.userId) {
+      return response.status(401).send()
+    }
     await prisma.memory.delete({ where: { id } })
   })
 }
